refactor(keyframes): extract DopesheetEditor marquee and timing-strip hooks
Pull two more state clusters out of DopesheetEditor: the timing-strip
drag handlers with their preview-frames state and ref mirror, and the
dopesheet marquee with its rectangle state, in-progress drag state,
click-dedup flag, hit-testing math, and global pointer listeners
(including the auto-scroll-on-edge behavior). Drops the orchestrator
from 3,191 to 3,028 LOC and removes the marquee constants and types
from the main file's import section. No behavior change.
